remove all shared hooks
